Change counter to u32

+++ src/counter.rs
@@ -1,19 +1,19 @@
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Counter {
-    counter: u16,
+    counter: u32,
 }
 
 impl Counter {
-    pub fn new(initial: u16) -> Self {
+    pub fn new(initial: u32) -> Self {
         Self { counter: initial }
     }
 
-    pub fn increment(&mut self) -> u16 {
+    pub fn increment(&mut self) -> u32 {
         self.counter += 1;
         self.counter
     }
 
-    pub fn value(&mut self) -> u16 {
+    pub fn value(&mut self) -> u32 {
         self.counter
     }
 }
